+++ src/routes/index.tsx
@@ -1,45 +1,10 @@
-/* eslint-disable @typescript-eslint/prefer-nullish-coalescing */
-import React, { useEffect, useState, useContext, useRef } from 'react'
-import {
-  BrowserRouter,
-  Redirect,
-  Route,
-  RouteProps,
-  Switch,
-  useHistory,
-  useLocation,
-} from 'react-router-dom'
-import { I18nextProvider, useTranslation } from 'react-i18next'
-import { Account } from '../views/Account'
-import { Login } from '../views/Login'
-import { NFT } from '../views/NFT'
-import { NFTs } from '../views/NFTs'
+import React, { useEffect } from 'react'
+import { BrowserRouter, Redirect, Route, Switch } from 'react-router-dom'
+import { I18nextProvider } from 'react-i18next'
 import { NotFound } from '../views/NotFound'
-import { Transfer } from '../views/Transfer'
 import i18n from '../i18n'
-import { Profile } from '../views/Profile'
-import { ImagePreview } from '../views/Profile/ImagePreview'
-import { TakePhoto } from '../views/Profile/TakePhoto'
-import { Explore } from '../views/Explore'
 import { Comfirm } from '../components/Confirm'
-import { useGetAndSetAuth, useProfile } from '../hooks/useProfile'
-import { Help } from '../views/Help'
-import { Unipass } from '../views/Unipass'
-import { Apps } from '../views/Apps'
-import { AddressCollector } from '../views/AddressCollector'
-import { useToast } from '../hooks/useToast'
-import { Collection } from '../views/Collection'
-import { Claim } from '../views/Claim'
-import { Issuer } from '../views/Issuer'
-import { UnipassConfig } from '../utils'
-import { Shop } from '../views/Shop'
-import { Redeem } from '../views/Reedem'
-import { RedeemDetail } from '../views/RedeemDetail'
-import { HolderAddress } from '../views/HolderAddress'
 import { WarningDialog } from '../components/WarningDialog'
-import { MyRedeem } from '../views/Reedem/My'
-import { RedeemPrize } from '../views/RedeemPrize'
-import { RedeemResult } from '../views/RedeemResult'
 import { ErrorToastDialog } from '../components/ErrorToast'
 import { GlobalSnackbar } from '../components/GlobalSnackbar'
 import { ConfirmToast } from '../components/ConfirmToast'
@@ -49,352 +14,11 @@ import {
   useLogin,
   WalletType,
 } from '../hooks/useAccount'
-
-export enum RoutePath {
-  Launch = '/',
-  Login = '/login',
-  Account = '/account',
-  NFT = '/nft/:id',
-  TokenClass = '/class/:id',
-  NFTs = '/home',
-  NotFound = '/404',
-  Transfer = '/transfer/:id',
-  Info = '/account/info',
-  Transactions = '/account/tx',
-  Profile = '/profile',
-  ImagePreview = '/avatar/preview',
-  TakePhoto = '/avatar/camera',
-  Explore = '/explore',
-  Help = '/help',
-  Unipass = '/unipass',
-  Apps = '/apps',
-  License = '/license',
-  Shop = '/shop',
-  AddressCollector = '/addresses',
-  Claim = '/claim',
-  Collection = '/explore/collection',
-  Issuer = '/issuer',
-  Redeem = '/redeem',
-  MyRedeem = '/redemption',
-  RedeemPrize = '/redeem-prize',
-  RedeemResult = '/redeem-result',
-  Holder = '/holder',
-  HolderAddress = '/holder/address',
-}
-
-export const RouterContext = React.createContext({
-  to: '',
-  from: '',
-})
-
-export interface Routes {
-  from: string
-  to: string
-}
-
-export const useRoute = (): Routes => {
-  return useContext(RouterContext)
-}
-
-const RouterProvider: React.FC = ({ children }) => {
-  const location = useLocation()
-  const [route, setRoute] = useState<Routes>({
-    to: location.pathname,
-    from: location.pathname,
-  })
-
-  useEffect(() => {
-    setRoute((prev) => ({ to: location.pathname, from: prev.to }))
-  }, [location])
-
-  return (
-    <RouterContext.Provider value={route}>{children}</RouterContext.Provider>
-  )
-}
-
-const allowWithoutAuthList = new Set([
-  RoutePath.Unipass,
-  RoutePath.Explore,
-  RoutePath.Apps,
-  RoutePath.AddressCollector,
-  RoutePath.Claim,
-  RoutePath.NotFound,
-  RoutePath.Redeem,
-  RoutePath.RedeemPrize,
-  RoutePath.RedeemResult,
-  RoutePath.MyRedeem,
-  RoutePath.Login,
-])
-
-const forceAuthList = new Set([`${RoutePath.Explore}?tag=follow`])
-
-const WalletChange: React.FC = ({ children }) => {
-  const { address, walletType, pubkey } = useAccount()
-  const { prevAddress, isLogined } = useAccountStatus()
-  const history = useHistory()
-  const location = useLocation()
-  useEffect(() => {
-    if (
-      prevAddress &&
-      address &&
-      prevAddress !== address &&
-      walletType &&
-      walletType !== WalletType.Unipass
-    ) {
-      history.push(RoutePath.NFTs)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [prevAddress, address, walletType])
-  const { isAuthenticated } = useProfile()
-  const getAuth = useGetAndSetAuth()
-  const isSigning = useRef(false)
-  const { toast, closeToast } = useToast()
-  const [t] = useTranslation('translations')
-
-  useEffect(() => {
-    const pathInForceAuthList = forceAuthList.has(
-      location.pathname + location.search
-    )
-    const pathInAllowList = [...allowWithoutAuthList].some((p) =>
-      location.pathname.startsWith(p)
-    )
-    if (
-      isLogined &&
-      !isAuthenticated &&
-      (!pathInAllowList || pathInForceAuthList) &&
-      !isSigning.current
-    ) {
-      if (
-        (WalletType.Unipass === walletType && pubkey) ||
-        WalletType.Metamask === walletType
-      ) {
-        isSigning.current = true
-        toast({
-          title: t('auth.title'),
-          content: t('auth.content'),
-          okText: t('auth.ok'),
-          showCloseIcon: false,
-          show: true,
-          onConfirm: () => {
-            if (pathInForceAuthList && WalletType.Unipass === walletType) {
-              UnipassConfig.setRedirectUri(location.pathname + location.search)
-            }
-            getAuth()
-              .then(() => {
-                if (WalletType.Metamask === walletType) {
-                  closeToast()
-                }
-              })
-              .catch(Boolean)
-          },
-        })
-      }
-    }
-  }, [
-    isAuthenticated,
-    walletType,
-    address,
-    location.pathname,
-    location.search,
-    isLogined,
-    pubkey,
-    t,
-    toast,
-    closeToast,
-    getAuth,
-  ])
-
-  return <>{children}</>
-}
-
-interface MibaoRouterProps extends RouteProps {
-  key: string
-  params?: string
-  path: string
-}
-
-const routes: MibaoRouterProps[] = [
-  {
-    component: NFTs,
-    exact: false,
-    key: 'NFTs',
-    path: RoutePath.NFTs,
-  },
-  {
-    component: Account,
-    exact: false,
-    key: 'Account',
-    path: RoutePath.Account,
-  },
-  {
-    component: NFT,
-    exact: true,
-    key: 'NFT',
-    path: RoutePath.NFT,
-  },
-  {
-    component: NFT,
-    exact: true,
-    key: 'TokenClass',
-    path: RoutePath.TokenClass,
-  },
-  {
-    component: Transfer,
-    exact: true,
-    key: 'Transfer',
-    path: RoutePath.Transfer,
-  },
-  {
-    component: Login,
-    exact: true,
-    key: 'Login',
-    path: RoutePath.Login,
-  },
-  {
-    component: Profile,
-    exact: false,
-    key: 'Profile',
-    path: RoutePath.Profile,
-  },
-  {
-    component: ImagePreview,
-    exact: true,
-    key: 'ImagePreview',
-    path: RoutePath.ImagePreview,
-  },
-  {
-    component: TakePhoto,
-    exact: true,
-    key: 'TakePhoto',
-    path: RoutePath.TakePhoto,
-  },
-  {
-    component: AddressCollector,
-    exact: true,
-    key: 'Addresses',
-    path: RoutePath.AddressCollector,
-    params: '/:id',
-  },
-  {
-    component: Explore,
-    exact: true,
-    key: 'Explore',
-    path: RoutePath.Explore,
-  },
-  {
-    component: Shop,
-    exact: true,
-    key: 'Shop',
-    path: RoutePath.Shop,
-  },
-  {
-    component: Help,
-    exact: false,
-    key: 'Help',
-    path: RoutePath.Help,
-  },
-  {
-    component: Help,
-    exact: false,
-    key: 'License',
-    path: RoutePath.License,
-  },
-  {
-    component: Unipass,
-    exact: false,
-    key: 'Unipass',
-    path: RoutePath.Unipass,
-  },
-  {
-    component: Apps,
-    exact: true,
-    key: 'Apps',
-    path: RoutePath.Apps,
-  },
-  {
-    component: Claim,
-    exact: true,
-    key: 'claim',
-    path: RoutePath.Claim,
-  },
-  {
-    component: Claim,
-    exact: true,
-    key: 'claim-with-id',
-    path: RoutePath.Claim,
-    params: '/:id',
-  },
-  {
-    component: Collection,
-    exact: true,
-    key: 'Collection',
-    path: RoutePath.Collection,
-    params: '/:id',
-  },
-  {
-    component: Redeem,
-    exact: true,
-    key: 'Redeem',
-    path: RoutePath.Redeem,
-  },
-  {
-    component: MyRedeem,
-    exact: true,
-    key: 'MyRedeem',
-    path: RoutePath.MyRedeem,
-  },
-  {
-    component: RedeemDetail,
-    exact: false,
-    key: 'ReedemDetail',
-    path: RoutePath.Redeem,
-    params: '/:id',
-  },
-  {
-    component: RedeemPrize,
-    exact: true,
-    key: 'ReedemPrize',
-    path: RoutePath.RedeemPrize,
-    params: '/:id',
-  },
-  {
-    component: RedeemResult,
-    exact: true,
-    key: 'ReedemResult',
-    path: RoutePath.RedeemResult,
-    params: '/:id',
-  },
-  {
-    component: Issuer,
-    exact: true,
-    key: 'Issuer',
-    path: RoutePath.Issuer,
-    params: '/:id',
-  },
-  {
-    component: HolderAddress,
-    exact: true,
-    key: 'HolderAddress',
-    path: RoutePath.HolderAddress,
-    params: '/:address',
-  },
-  {
-    component: NFTs,
-    exact: true,
-    key: 'Holder',
-    path: RoutePath.Holder,
-    params: '/:address',
-  },
-]
-
-export enum ProfilePath {
-  Regions = '/profile/regions',
-  Provinces = '/profile/regions/provinces',
-  Cities = '/profile/regions/cities',
-  Username = '/profile/username',
-  Description = '/profile/description',
-  Birthday = '/profile/birthday',
-}
+import { RoutePath } from './path'
+import { AccountChange } from './AccountChange'
+import { routes } from './routes'
+import { RouterProvider } from '../hooks/useRoute'
+export * from './path'
 
 export const Routers: React.FC = () => {
   const { walletType } = useAccount()
@@ -413,7 +37,7 @@ export const Routers: React.FC = () => {
     <I18nextProvider i18n={i18n}>
       <BrowserRouter>
         <RouterProvider>
-          <WalletChange>
+          <AccountChange>
             <Switch>
               {routes.map((route) => (
                 <Route
@@ -439,7 +63,7 @@ export const Routers: React.FC = () => {
             <ConfirmToast />
             <GlobalSnackbar />
             <Comfirm open disableBackdropClick />
-          </WalletChange>
+          </AccountChange>
         </RouterProvider>
       </BrowserRouter>
     </I18nextProvider>
